scripts/sdxl/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
diffusers==0.27.2
transformers==4.39.3
optimum-intel[openvino]==1.17.0
huggingface_hub[cli]
